Create PVE admin users
Introduce the role variable proxmox_pveadmins where - when defined - the
PVE group 'admin' is created and the ACL is updated so that members of
the group 'admin' are assigned the 'Administrator' role for the whole
system (path '/').

Users, groups and roles/ACLs are controlled with the command pveum(1)
(Proxmox VE User Manager) but fails if an user, group or role/ACL
already exists, which makes the command not idempodent.

Unfortunately 'pveum' also lacks a 'list' command (for users, groups
and roles/ACLs) so pvesh(1) (Shell interface for the Proxmox VE API) and
'/etc/pve/user.cfg' is used to check if the user, group and role/ACL exists.

See:

* https://forum.proxmox.com/threads/how-to-list-user-permisions.49425/
* https://bugzilla.proxmox.com/show_bug.cgi?id=1502