Fix ansible-lint warnings
Running "ansible-lint --offline ." caused some warnings.
Removed .yamllint because it was created a while ago via molecule init,
but we currently do not use molecule for this role.

The namespace "synpro" is not final, it depends on what is possible in
case we move this role to the SynProSolutions repositories.