Add Ceph No-Subscription Repository
We now deploy `/etc/apt/sources.list.d/ceph.list with
the Ceph No-Subscription Repository.

We default to "ceph-squid" which is the current latest support version
of Proxmox VE 8.4. For older versions "proxmox_ceph_version" needs to be
set in group_vars for the cluster.

Remove the examples for the `clusternode.ceph` entries in /etc/hosts.
They are not needed and should be removed.